+++ lib/contentful.js
@@ -1,4 +1,5 @@
-import { dateString } from '~/lib/date'
+// nuxt.config.jsからも呼び出すため相対パスにする
+import { dateString, getAllTerm } from './date'
 const contentful = require('contentful')
 
 // 初期設定
@@ -10,10 +11,37 @@ const client = contentful.createClient({
 })
 
 /*
+ * ルーティングのパス取得
+ * @return array ページパス一覧
+ */
+export function routing() {
+  const allTerm = getAllTerm()
+  return client
+    .getEntries({
+      content_type: contentType
+    })
+    .then(entries => {
+      const page = Math.ceil(entries.total / process.env.constant.postsPerPage)
+      return [
+        ...entries.items.map(entry => `/posts/${entry.fields.url}/`),
+        // ページング
+        ...[...Array(page).keys()].map(i => `/posts/page/${i + 1}/`),
+        // 日次のパス生成
+        ...allTerm.map(data => `/news/${data.key}/${data.value}/`),
+        // 月次のパス生成
+        ...allTerm.map(data => `/news/${data.key}/`)
+      ]
+    })
+    .catch(error => {
+      console.log(error)
+    })
+}
+/*
  * 記事の取得
  * @param limit 取得件数
  * @param skip 取得開始位置
  */
+
 export async function getEntries(limit, page = 1) {
   return client
     .getEntries({
