+++ nuxt.config.js
@@ -1,5 +1,5 @@
 import VuetifyLoaderPlugin from 'vuetify-loader/lib/plugin'
-import { getAllTerm } from './lib/date'
+import { routing } from './lib/contentful'
 import constant from './config/constant.json'
 const environment = process.env.NODE_ENV || 'dev'
 const conf = require(`./config/constant.${environment}.json`)
@@ -16,40 +16,6 @@ const storageBucket = process.env.STORAGE_BUCKET || 'storageBucket'
 const messagingSenderId = process.env.MESSAGING_SENDER_ID || 'messagingSenderId'
 const appId = process.env.APP_ID || 'appId'
 const adsenseId = process.env.ADSENSE_ID || 'adsenseId'
-
-// contentfulから記事を取得する
-const contentful = require('contentful')
-const client = contentful.createClient({
-  space: space,
-  accessToken: accessToken,
-  timeout: 60000,
-  retryLimit: 10
-})
-
-// ルーティング
-function routing() {
-  const allTerm = getAllTerm()
-  return client
-    .getEntries({
-      content_type: contentType
-    })
-    .then(entries => {
-      const page = Math.ceil(entries.total / constant.postsPerPage)
-      return [
-        ...entries.items.map(entry => `/posts/${entry.fields.url}/`),
-        // ページング
-        ...[...Array(page).keys()].map(i => `/posts/page/${i + 1}/`),
-        // 日次のパス生成
-        ...allTerm.map(data => `/news/${data.key}/${data.value}/`),
-        // 月次のパス生成
-        ...allTerm.map(data => `/news/${data.key}/`)
-      ]
-    })
-    .catch(error => {
-      console.log('error')
-      console.log(error)
-    })
-}
 
 export default {
   // 本来spaモードだがそれだとページごとのmetaが反映されないので
