+++ pages/news/index.vue
@@ -29,7 +29,6 @@
 
 <script>
 import { mapActions, mapGetters } from 'vuex'
-import { getEntries } from '~/plugins/contentful'
 import { dateString } from '~/lib/date'
 import CardItem from '~/components/organisms/CardItem.vue'
 import ListItem from '~/components/organisms/ListItem.vue'
