+++ pages/posts/page/_number.vue
@@ -24,7 +24,7 @@
 
 <script>
 import { mapActions, mapGetters } from 'vuex'
-import { getEntries } from '~/plugins/contentful'
+import { getEntries } from '~/lib/contentful'
 import { dateString } from '~/lib/date'
 import { checkPage } from '~/lib/validation'
 import ListItem from '~/components/organisms/ListItem.vue'
